refactor(mkdir): consistent error handling
This refactors mkdir() to (for the most part) invoke mkdirSync() and
catch any exceptions, instead of the error-prone approach of trying to
predict when specific errors would happen.

This adds two new tests for "name too long" and that it's OK to `mkdir
-p` a directory which already exists. The latter was implicitly covered
by another test case, but this should be explicit.

Lastly, this changes error messages to be more consistent with Linux
mkdir, and adds the "avoid-escape" rule in the test directory to allow
using double-quoted strings when it aids readability.

Fixes #722