test: add more coverage for globOptions
No change to logic. This adds tests for more `config.globOptions`
values. This aims to cover the options which appear to be supported by
both node-glob and fast-glob, however the options have different names
in each module. This is a followup to PR #1163.

See https://github.com/mrmlnc/fast-glob?tab=readme-ov-file#compatible-with-node-glob

Issue #828