chore: update ava dep
No change to logic. This updates ava to major version 2. This is not the
latest version, however updating to version 3 is not trivial because of
some breaking changes.

This removes the commandline options from the ava invocation in favor of
moving this into the well-supported ava JSON configuration. This is
required for Windows compatibility, because otherwise Windows won't
properly expand "test/*.js" on the commandline and ava v2 won't expand
this unless it's specified in the JSON config.

Test: npm test
Test: npm run test-with-coverage