refactor: avoid dependency cycle
Prior to this change, there was a dependency cycle between shell.js and
common.js. This refactors code to avoid the cycle.

This was previously reported in #1133. Node.js already worked around the
issue in v20.6.1, but it's simpler if we can avoid the dependency cycle
on our end anyway.

I tested this with `npm test` and also by running my own script which
uses a shelljs plugin. This seems to work correctly in both cases.