refactor: use es6 class syntax for CommandError
This is supposed to be a NOOP, although for some reason this makes a
difference in the ava@6 test framework. This should not have a
user-visible effect because CommandError is an implementation detail of
this library (it's only thrown if `options.fatal` is off, and it's
always caught inside of common.js).